Animations, collision mask, central player systems
Updated the animations to correctly use collision masks with the new setup.

Updated component responsibility in regards to animations and the state machine. ex: input is emitted, a component takes that emission, filters it through the relevant conditions, and finally emits again for the animation manager to handle. Code is now more compact and less spaghetti and components are *somewhat* better about relevant responsibility.

Added a box and a shadow. :)